+++ jieunLee/spring/bank/src/main/java/com/example/bank/BankApplication.java
@@ -1,6 +1,6 @@
 package com.example.bank;
 
-import com.example.bank.controller.bank9.Bank7Controller;
+import com.example.bank.controller.bank7.Bank7Controller;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
