--- jieunLee/spring/bank/src/main/java/com/example/bank/controller/ProductController.java
@@ -1,76 +1,0 @@
-package com.example.bank.controller;
-
-import com.example.bank.entity.ProductBoard;
-import com.example.bank.service.ProductBoardService;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Slf4j
-@Controller
-@RequestMapping("/product")
-public class ProductController {
-
-    @Autowired
-    private ProductBoardService productBoardService;
-
-    @GetMapping("/register")
-    public String registerForm (ProductBoard productBoard, Model model) {
-        log.info("registerForm check");
-
-        return "/product/register";
-    }
-
-    @PostMapping("/register")
-    public String register (ProductBoard productBoard, Model model) {
-        log.info("register check");
-
-        productBoardService.register(productBoard);
-        return "/product/register";
-    }
-
-    @GetMapping("/list")
-    public String list (Model model) {
-        log.info("list check");
-
-        model.addAttribute("list", productBoardService.list());
-
-        return "/product/list";
-    }
-
-    @GetMapping("/read")
-    public String read (int productNo, Model model) {
-        log.info("read check");
-
-        model.addAttribute(productBoardService.read(productNo));
-        return "/product/read";
-    }
-
-    @GetMapping("/modify")
-    public String modifyForm (int productNo, Model model) {
-        log.info("modifyForm check");
-
-        model.addAttribute(productBoardService.read(productNo));
-        return "/product/modify";
-    }
-
-    @PostMapping("modify")
-    public String modify (ProductBoard productBoard, Model model) {
-        log.info("modify check");
-
-        productBoardService.modify(productBoard);
-        return "/product/success";
-    }
-
-    @PostMapping("/remove")
-    public String remove (int productNo, Model model) {
-        log.info("remove");
-
-        productBoardService.remove(productNo);
-        return "/product/success";
-    }
-}
