--- jieunLee/spring/bank/src/main/java/com/example/bank/entity/Board.java
@@ -1,14 +1,0 @@
-package com.example.bank.entity;
-
-import lombok.Data;
-
-import java.util.Date;
-
-@Data
-public class Board {
-    private int boardNo;
-    private String title;
-    private String content;
-    private String writer;
-    private Date regDate;
-}
