--- jieunLee/spring/bank/src/main/java/com/example/bank/entity/ViewMyLotto.java
@@ -1,14 +1,0 @@
-package com.example.bank.entity;
-
-import lombok.Data;
-
-@Data
-public class ViewMyLotto {
-    private int myNum;
-    private String yesOrNo;
-
-    public ViewMyLotto(int myNum, String yesOrNo) {
-        this.myNum = myNum;
-        this.yesOrNo = yesOrNo;
-    }
-}
