--- jieunLee/spring/bank/src/main/java/com/example/bank/entity/ViewRandomProduct.java
@@ -1,12 +1,0 @@
-package com.example.bank.entity;
-
-import lombok.Data;
-
-@Data
-public class ViewRandomProduct {
-    private Object[] randPro;
-
-    public ViewRandomProduct(Object[] randPro) {
-        this.randPro = randPro;
-    }
-}
