+++ jieunLee/spring/bank/src/main/java/com/example/bank/entity/bank9/ViewLottoPlacement.java
@@ -1,0 +1,14 @@
+package com.example.bank.entity.bank9;
+
+import lombok.Data;
+
+@Data
+public class ViewLottoPlacement {
+    private int[] totalArr;
+    private int[] lottoArr;
+
+    public ViewLottoPlacement(int[] totalArr, int[] lottoArr) {
+        this.totalArr = totalArr;
+        this.lottoArr = lottoArr;
+    }
+}
