+++ jieunLee/spring/bank/src/main/java/com/example/bank/entity/bank9/ViewRandomProduct.java
@@ -1,0 +1,12 @@
+package com.example.bank.entity.bank9;
+
+import lombok.Data;
+
+@Data
+public class ViewRandomProduct {
+    private Object[] randPro;
+
+    public ViewRandomProduct(Object[] randPro) {
+        this.randPro = randPro;
+    }
+}
