+++ jieunLee/spring/bank/src/main/java/com/example/bank/entity/order32/Board.java
@@ -1,0 +1,14 @@
+package com.example.bank.entity.order32;
+
+import lombok.Data;
+
+import java.util.Date;
+
+@Data
+public class Board {
+    private int boardNo;
+    private String title;
+    private String content;
+    private String writer;
+    private Date regDate;
+}
