+++ jieunLee/spring/bank/src/main/java/com/example/bank/service/order32/BoardServiceImpl.java
@@ -1,0 +1,40 @@
+package com.example.bank.service.order32;
+
+import com.example.bank.entity.order32.Board;
+import com.example.bank.repository.order32.BoardRepository;
+import com.example.bank.service.order32.BoardService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class BoardServiceImpl implements BoardService {
+    @Autowired
+    BoardRepository repository;
+
+    @Override
+    public void register(Board board) {
+        repository.create(board);
+    }
+
+    @Override
+    public List<Board> list() {
+        return repository.list();
+    }
+
+    @Override
+    public Board read(Integer boardNo) {
+        return repository.read(boardNo);
+    }
+
+    @Override
+    public void modify(Board board) {
+        repository.update(board);
+    }
+
+    @Override
+    public void remove(Integer boardNo) {
+        repository.delete(boardNo);
+    }
+}
