+++ jieunLee/spring/bank/src/main/java/com/example/bank/utility/bank8/Sequence.java
@@ -1,0 +1,27 @@
+package com.example.bank.utility.bank8;
+
+import lombok.Data;
+
+import java.math.BigInteger;
+
+@Data
+public class Sequence {
+    final BigInteger INIT = new BigInteger("1");
+    final BigInteger BASE = new BigInteger("2");
+
+    BigInteger[] seq;
+    int length;
+
+    public Sequence(int idx) {
+        length = idx;
+    }
+
+    public BigInteger initSeq() {
+        seq = new BigInteger[length];
+        seq[0] = INIT;
+        for(int i=1; i<length; i++) {
+            seq[i] = seq[i-1].multiply(BASE);
+        }
+        return seq[length-1];
+    }
+}
