+++ jieunLee/spring/bank/src/main/java/com/example/bank/utility/bank9/ViewEquTriangleArea.java
@@ -1,0 +1,14 @@
+package com.example.bank.utility.bank9;
+
+import lombok.Data;
+
+@Data
+public class ViewEquTriangleArea {
+    private int base;
+    private double area;
+
+    public ViewEquTriangleArea(int base, double area) {
+        this.base = base;
+        this.area = area;
+    }
+}
