+++ jieunLee/spring/bank/src/main/java/com/example/bank/utility/bank9/ViewRegHexagonArea.java
@@ -1,0 +1,14 @@
+package com.example.bank.utility.bank9;
+
+import lombok.Data;
+
+@Data
+public class ViewRegHexagonArea {
+    private int base;
+    private double area;
+
+    public ViewRegHexagonArea(int base, double area) {
+        this.base = base;
+        this.area = area;
+    }
+}
