--- jieunLee/spring/bank/src/main/resources/templates/product/register.html
@@ -1,61 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.w3.org/1999/xhtml">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-
-<script type="text/javascript" src="http://code.jquery.com/jquery-latest.min.js"></script>
-<script>
-    $(document).ready(function() {
-        var formObj = $("#productBoard");
-
-        $("#btnRegister").on("click", function() {
-            formObj.attr("action", "/product/register");
-            formObj.attr("method", "post");
-            formObj.submit();
-        });
-
-        $("#btnList").on("click", function() {
-            self.location = "/product/list";
-        });
-    });
-</script>
-
-<body>
-    <h2>글쓰기 페이지</h2>
-    <form id="productBoard" th:object="${productBoard}">
-        <table>
-            <tr>
-                <td>제목</td>
-                <td>
-                    <input type="text" name="title" th:field="*{title}"/>
-                </td>
-            </tr>
-            <tr>
-                <td>작성자</td>
-                <td>
-                    <input type="text" name="writer" th:field="*{writer}"/>
-                </td>
-            </tr>
-            <tr>
-                <td>본문</td>
-                <td>
-                    <input type="text" name="description" th:field="*{description}"/>
-                </td>
-            </tr>
-            <tr>
-                <td>가격</td>
-                <td>
-                    <input type="text" name="price" th:field="*{price}"/>
-                </td>
-            </tr>
-        </table>
-    </form>
-
-    <div>
-        <button type="submit" id="btnRegister">등록</button>
-        <button type="submit" id="btnList">돌아가기</button>
-    </div>
-</body>
-</html>
